refactor: jwt annotation 명 수정
#186
