+++ backend/src/main/java/sw_css/utils/JwtToken/SuperAdminArgumentResolver.java
@@ -15,7 +15,7 @@ import sw_css.member.exception.MemberException;
 import sw_css.member.exception.MemberExceptionType;
 import sw_css.utils.JwtToken.exception.JwtTokenException;
 import sw_css.utils.JwtToken.exception.JwtTokenExceptionType;
-import sw_css.utils.annotation.Admin;
+import sw_css.utils.annotation.SuperAdminInterface;
 
 @Component
 @RequiredArgsConstructor
@@ -26,7 +26,7 @@ public class SuperAdminArgumentResolver implements HandlerMethodArgumentResolver
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(Admin.class);
+        return parameter.hasParameterAnnotation(SuperAdminInterface.class);
     }
 
     @Override
