+++ backend/src/main/java/sw_css/utils/annotation/AdminInterface.java
@@ -7,6 +7,6 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Admin {
+public @interface AdminInterface {
     boolean required() default true;
 }
